feat: 로또 당첨번호 입력 기능 작성
