+++ src/main/java/lotto/controller/LottoController.java
@@ -1,10 +1,13 @@
 package lotto.controller;
 
-import lotto.utils.InputView;
+import java.util.List;
+import lotto.view.InputView;
 
 public class LottoController {
 
     public void runLotto() {
         int lottoPurchaseAmount = InputView.inputLottoPurchaseAmount();
+        String wonLottoNumbersString = InputView.inputWonLottoNumbers();
+        List<Integer> parsedWonLottoNumbers = InputView.parseWonLottoNumbers(wonLottoNumbersString);
     }
 }
